Optimized queries, changed a little bit the database (with maybe a bug with nocturnal bugs, damnit ACTV!)
It seems like night buses which still rely to the present day have the following hours in timetables:
24 -> midnight
25 -> one am
26 -> two am

and so on...

BUT

if I change 25 with 01 what I get is that a bus which would be - let's say - late night on friday evening would be then moved to early friday morning. I'd like to be able to move it to early SATURDAY morning, but it's not so easy.